adapt livegain plugins for the new host
