+++ packages/host/src/index.html
@@ -24,7 +24,10 @@
 					<li data-plugin-url="quadrafuzz_without_builder/src">Quadrafuzz without builder</li>
 					<li data-plugin-url="pingpongdelay/dist">PingPongDelay</li>
 					<li data-plugin-url="faustPingPongDelay/plugin">faustPingPongDelay</li>
-					<li data-plugin-url="livegain/dist">LiveGain</li>
+					<li data-plugin-url="livegain/dist/livegain">LiveGain</li>
+					<li data-plugin-url="livegain/dist/oscilloscope">Oscilloscope</li>
+					<li data-plugin-url="livegain/dist/spectroscope">Spectroscope</li>
+					<li data-plugin-url="livegain/dist/spectrogram">Spectrogram</li>
 				</ul>
 			</div>
 			<div>
