+++ packages/livegain/src/livegain/LiveGainNode.ts
@@ -1,8 +1,8 @@
 /* eslint-disable import/no-duplicates */
 import { CompositeAudioNode, ParamMgrNode } from "sdk";
-import { TemporalAnalyserNode } from "./worklets/TemporalAnalyser";
-import { atodb, dbtoa } from "./utils/math";
-import { Parameters } from "./LiveGainModule";
+import { TemporalAnalyserNode } from "../worklets/TemporalAnalyser";
+import { atodb, dbtoa } from "../utils/math";
+import { Parameters } from ".";
 
 export default class LiveGainNode extends CompositeAudioNode {
     analyserNode: TemporalAnalyserNode;
