+++ packages/livegain/src/spectroscope/SpectroscopeUI.tsx
@@ -4,19 +4,16 @@
 /* eslint-disable react/sort-comp */
 /* eslint-disable react/destructuring-assignment */
 import * as React from "react";
-import { atodb } from "./utils/math";
-import type Module from "./SpectrogramModule";
+import Color from "color-js";
+import { atodb } from "../utils/math";
+import type Module from ".";
 
-export default class SpectrogramUI extends React.PureComponent<{ module: Module }> {
+export default class SpectroscopeUI extends React.PureComponent<{ module: Module }> {
     static defaultSize: [number, number] = [640, 360];
     refCanvas = React.createRef<HTMLCanvasElement>();
     paintScheduled = false;
     $paintRaf = -1;
     $changeTimer = -1;
-    $lastFrame = -1;
-    frames = 1;
-    offscreenCtx = document.createElement("canvas").getContext("2d");
-    offscreenVRes = 1024;
     get canvas() {
         return this.refCanvas.current;
     }
@@ -24,7 +21,7 @@ export default class SpectrogramUI extends React.PureComponent<{ module: Module 
         return this.refCanvas.current ? this.refCanvas.current.getContext("2d") : null;
     }
     fullSize(): [number, number] {
-        const [width, height] = SpectrogramUI.defaultSize;
+        const [width, height] = SpectroscopeUI.defaultSize;
         const { canvas, ctx } = this;
         if (!ctx) return [0, 0];
         if (typeof width === "number" && typeof height === "number") {
@@ -57,9 +54,6 @@ export default class SpectrogramUI extends React.PureComponent<{ module: Module 
         this.paintScheduled = true;
     };
     componentDidMount() {
-        const { offscreenCtx, frames } = this;
-        offscreenCtx.canvas.width = frames;
-        offscreenCtx.canvas.height = this.offscreenVRes;
         this.schedulePaint();
     }
     componentDidUpdate() {
@@ -70,69 +64,31 @@ export default class SpectrogramUI extends React.PureComponent<{ module: Module 
     }
     async paint() {
         const bgColor = "rgb(40, 40, 40)";
+        const fgColor = "hsl(0, 100%, 85%)";
+        const hueOffset = 60;
         const gridColor = "#404040";
         const seperatorColor = "white";
-        const { ctx, offscreenCtx, offscreenVRes } = this;
-        if (!ctx || !offscreenCtx) return;
+        const ctx = this.ctx;
+        if (!ctx) return;
 
 
         const left = 0;
         const bottom = 0;
 
-        const allAmplitudes = await this.props.module.audioNode.analyserNode.getAllAmplitudes();
+        const lastAmplitudes = await this.props.module.audioNode.analyserNode.getLastAmplitudes();
 
         // Background
-
         const [width, height] = this.fullSize();
         ctx.clearRect(0, 0, width, height);
         ctx.fillStyle = bgColor;
         ctx.fillRect(0, 0, width, height);
 
-        if (!allAmplitudes) return;
-        const { data: f, $totalFrames, fftBins: bins, frames: framesIn, $frame: $frameUi32 } = allAmplitudes;
+        if (!lastAmplitudes) return;
+        const { data: f } = lastAmplitudes;
         if (!f || !f.length || !f[0].length) return;
         const l = f[0].length;
         const channels = f.length;
 
-        // Draw to offscreen canvas
-        let frames = this.frames;
-        const $lastFrame = $totalFrames[0] - 1;
-        const $frame = $frameUi32[0];
-        let $frame0 = $frame;
-        const $frame1 = $frame0 + framesIn;
-        if (frames !== framesIn) {
-            offscreenCtx.canvas.width = framesIn;
-            this.frames = framesIn;
-            frames = framesIn;
-        } else if ($lastFrame >= this.$lastFrame) {
-            $frame0 = Math.max($frame0, $frame1 - ($lastFrame - this.$lastFrame));
-        }
-        this.$lastFrame = $lastFrame;
-        const osChannelHeight = offscreenVRes / channels;
-        const step = Math.max(1, Math.round(bins / osChannelHeight));
-        const vGrid = osChannelHeight / bins;
-        for (let i = 0; i < f.length; i++) {
-            for (let j = $frame0; j < $frame1; j++) {
-                let maxInStep;
-                offscreenCtx.fillStyle = "black";
-                offscreenCtx.fillRect(j % frames, i * osChannelHeight, 1, osChannelHeight);
-                for (let k = 0; k < bins; k++) {
-                    const samp = atodb(f[i][(k + j * bins) % l]);
-                    const $step = k % step;
-                    if ($step === 0) maxInStep = samp;
-                    if ($step !== step - 1) {
-                        if ($step !== 0 && samp > maxInStep) maxInStep = samp;
-                        continue;
-                    }
-                    const normalized = Math.min(1, Math.max(0, (maxInStep + 10) / 100 + 1));
-                    if (normalized === 0) continue;
-                    const hue = (normalized * 180 + 240) % 360;
-                    const lum = normalized * 50;
-                    offscreenCtx.fillStyle = `hsl(${hue}, 100%, ${lum}%)`;
-                    offscreenCtx.fillRect(j % frames, (bins - k - 1) * vGrid + i * osChannelHeight, 1, Math.max(1, vGrid));
-                }
-            }
-        }
         // Grids
         ctx.strokeStyle = gridColor;
         const vStep = 0.25;
@@ -164,20 +120,36 @@ export default class SpectrogramUI extends React.PureComponent<{ module: Module 
         }
         ctx.stroke();
         ctx.setLineDash([]);
+        ctx.lineWidth = 2;
+        const channelColor: string[] = [];
         // Horizontal Range
-        ctx.save();
-        ctx.globalCompositeOperation = "lighter";
-        ctx.imageSmoothingEnabled = false;
-        $frame0 = $frame;
-        if ($frame1 === frames) {
-            ctx.drawImage(offscreenCtx.canvas, 0, 0, frames, offscreenVRes, left, 0, width - left, height - bottom);
-        } else {
-            const sSplit = frames - $frame0;
-            const dSplit = sSplit / frames * (width - left);
-            ctx.drawImage(offscreenCtx.canvas, $frame0, 0, sSplit, offscreenVRes, left, 0, dSplit, height - bottom);
-            ctx.drawImage(offscreenCtx.canvas, 0, 0, $frame1 - frames - 0.01, offscreenVRes, dSplit + left, 0, width - left - dSplit, height - bottom);
+        const $0 = 0; // Draw start
+        const $1 = l; // Draw End
+        const gridX = (width - left) / ($1 - $0);
+        const step = Math.max(1, Math.round(1 / gridX));
+        for (let i = 0; i < f.length; i++) {
+            ctx.beginPath();
+            channelColor[i] = Color(fgColor).shiftHue(i * hueOffset).toHSL();
+            ctx.fillStyle = channelColor[i];
+            let maxInStep;
+            for (let j = $0; j < $1; j++) {
+                const samp = atodb(f[i][j]);
+                const $step = (j - $0) % step;
+                if ($step === 0) maxInStep = samp;
+                if ($step !== step - 1) {
+                    if ($step !== 0 && samp > maxInStep) maxInStep = samp;
+                    continue;
+                }
+                const x = (j - $0) * gridX + left;
+                const y = channelHeight * (i + 1 - Math.min(1, Math.max(0, maxInStep / 100 + 1)));
+                if (j === $0) ctx.moveTo(x, y);
+                else ctx.lineTo(x, y);
+            }
+            ctx.lineTo(width, channelHeight * (i + 1));
+            ctx.lineTo(left, channelHeight * (i + 1));
+            ctx.closePath();
+            ctx.fill();
         }
-        ctx.restore();
     }
     render() {
         const defaultCanvasStyle: React.CSSProperties = { position: "relative", display: "inline-block", backgroundColor: "white" };
